Segmentation/requirements.txt
# requirements
# Usage: pip install -r requirements.txt

numpy>=1.18.5
scipy>=1.7.3
torch>=1.10.1  # see https://pytorch.org/get-started/locally (recommended)
torchvision>=0.11.2
pandas>=1.4.1
kornia>=0.6.3
